move all mac depoly relates into mac_build
